Handle timestamp rewriting bug
If the CLAN interval that needs to be adjusted happens to be really
short, then a timestamp adjust runs the risk of reseting the offset to
come prior to onset. When first inserting the silences, these timestamp
rewrites can become malformed for subsequent steps (subregion insert).

We decided the easiest way to take care of this would be to print an
error message indicating where the problem was, and manually going and
adjusting the interval so that CLAN can read it. The main problem here
is that altering the data to begin with was a bad idea. Coming up with
special forms of adjusting timestamps for every edge case becomes
unmanagable pretty quickly. This fixes #1, at least in the short term.
Spending time on a more general solution would be worthwhile.