Implement "Load All" feature
There's now a button called "Load All" which expects the path to the
main CLAN file (e.g 16_08.cex). It then loads and generates all the
other files that are necessary, running through all the steps at once.
It assumes that all the necessary files are within the same directory
as the original CLAN file that was loaded. It will output the
silences_added and subregions exports to this same directory. The old
functionality is still in place, so if you want to go and load/export
everything one at a time, you can still do that.

The format it's expecting files to be in:

    16_08.cex
    16_08_silences.txt
    16_08_lena5min.csv
    16_08_silences_added.cex
    16_08_subregions.cex

There's also a new function that scans the entire subregions export
after it's been created to check for onsets that come after offsets. It
will print debug messages to the console to show where in the CLAN file
this occurs. These should be adjusted by hand so that CLAN can read in
the file. I'll try and remove this manual labor from the process as
soon as I can, but having a human adjust malformed timestamps is
probably a more reliable method than having the computer try to fix
errors that it itself created in previous steps of the pipeline.
Otherwise, the whole timestamp rewriting process spirals into edge case
handling.