Make sure filepaths and loading/generation are correct
The last 3 commits change the behavior of the program so that the "Load
All" button looks for a "Label_Track.txt" file in the directory of the
main CLAN file being loaded. It will then generate the
XX_YY_silences.txt file with regions of sound shorter than the minimum
sound length removed from the list. These new regions are the ones that
audiowords will work with.