Workaround single sound region bug
When there's a single sound region in the Label_Track.txt we end up
with no silence regions. Parts of the program depend on there being at
least one silence region. This commit sidesteps the issue by
artificially inserting a 2 second silence at the beginning. This isn't
an optimal solution and should be revisited. This takes care of the
problem with 25_08 shown in issue #5, but only goes halfway in fixing
the problem with 37_06. It won't throw an error while processing 37_06
anymore but it should figure out that the single sound region is surrounded
by silence on both sides. I'll close the issue when this is taken care of too.