Fix #7
This is a continuation of the fixes from the previous two commits. If
there's only one sound region (hence no silences), we need to be able
to run the script without silences. This adds a check for the condition
where Label_Track.txt has only 2 regions, one of them being the end
marker. In this case, we add an artificial silence in the beginning to
make sure the sript runs all the way through.