Working on new clan format handling
Started writing copies of the functions currently being used to parse
through the old clan format. The old functionality is going to stay in
place so that we can still use audiowords on .cex files. The new
functionality is encapsulated into the new "Load All (cha)" button. It
behaves the same way as "Load All" except it's expecting a .cha file.
This isn't fully implemented yet, but it's relatively close.