Fix counting error for silences
This will fix #10. The new parse_silences function was incrementing the
count of the silence even when the silence is thrown away. In this
case, it's because it was less than 15 milliseconds. This check was
introduced in the last commit to handle the case where the first sound
starts at 0.01 seconds, meaning there's silence between 0 and 0.01 seconds.